Fix bug in estep (dnorm asks for stdev, not variance)
E-step (for dimdat = 1) was using variance, not standard deviation, as an input
to `dnorm()`. This is fixed in this commit.

Gating is also updated so that the 1E-10 that is added to the weighted densities
(prior to forming the responsibilities) is customizeable and defaults to a
smaller value (1E-20); this is called `eps` in this function now.